+++ code/setup_and_dataloading.h
@@ -19,7 +19,7 @@
  **/
 namespace configld {
 
-    bool load_config_file(int scenario_id, std::string& filepath); ///< Load the config file, that is passed as command line argument
+    bool load_config_file(unsigned long scenario_id, std::string& filepath); ///< Load the config file, that is passed as command line argument
 
     bool load_data_from_central_database(const char* filepath); ///< Load the complete simulation structure from the central database, and also load the recorded time series for the measurement units
